requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
django-filter==2.4.0
django-guardian==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
psycopg2==2.9.1
Pygments==2.10.0
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
whitenoise==5.3.0
